+++ lib/broadside.rb
@@ -1,5 +1,5 @@
 require 'broadside/error'
-require 'broadside/utils'
+require 'broadside/logging_utils'
 require 'broadside/configuration/verify_instance_variables'
 require 'broadside/configuration'
 require 'broadside/configuration/aws_config'
@@ -12,7 +12,7 @@ require 'broadside/ecs/ecs_manager'
 require 'broadside/version'
 
 module Broadside
-  extend Utils
+  extend LoggingUtils
 
   USER_CONFIG_FILE = "#{Dir.home}/.broadside/config.rb"
 
@@ -45,5 +45,7 @@ module Broadside
 
   def self.reset!
     @config = nil
+    EcsManager.instance_variable_set(:@ecs_client, nil)
+    EcsManager.instance_variable_set(:@ec2_client, nil)
   end
 end