+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # 3.0.0
 - **BREAKING CHANGE**: `ssh`, `bash`, `logtail`, `status`, and `run` are now top level commands, not subcommands of `deploy`
-- **BREAKING CHANGE**: `config.git_repo` was removed.  `config.base` and `config.deploy` also are no longer backwards compatible
-- Put back per target `tag:` config
+- **BREAKING CHANGE**: `config.git_repo=` and `config.type=` were removed.
+- **BREAKING CHANGE**: `config.base` and `config.deploy` are no longer backwards compatible
+- **BREAKING CHANGE**: `instance` can no longer be configured on a per `Target` basis
+- *NEW FEATURE*: Allow per target `:docker_image` configuration
+- *NEW FEATURE*: Put back per target `:tag` configuration
+- *NEW FEATURE*: Add `list_targets` command to display all the targets' deployed images and CPU/memory allocations
 - Only load `env_files` for the selected target (AKA don't preload everything when you aren't using it)
-- Add `list_targets` command to display all the targets' deployed images and CPU/memory allocations
+- Make `env_files` configuration optional
+- No more `Utils` module, just `LoggingUtils`
+- Exceptions will be raised if you try to configure a target with an invalid hash key
 
 # 2.0.0
 - **BREAKING CHANGE**: `rake db:migrate` is no longer the default `predeploy_command`